refactor: move filesystem tree construct to backend Signed-off-by: Philippe Martin <phmartin@redhat.com>
