+++ packages/api/src/filesystem-tree.spec.ts
