+++ packages/api/src/filesystem-tree.ts
